Tarball compress/decompress
Add tarball compress/decompress function: This bash script prompts the user for an operation - either compress ('c') or decompress ('d'). Depending on the user's choice, it asks for further inputs like the directory of files to be compressed or the tarball to be decompressed and carries out the specified operation.